refactor: extract wireEvents into js/event-wiring.js
Extract wireEvents (~1,460 lines), handleKeyboard (~197 lines), and
selectTool (~78 lines) from app.js into js/event-wiring.js. Feature
handlers that remain in app.js are accessed via a callback bridge
(setEventCallbacks). Module-level variables like cropState,
_selectedLinkObj, etc. are accessed through getter callbacks.

app.js reduced from ~5,540 to ~3,840 lines. All 621 tests pass.
